Create install_arch.sh
I am providing a script to install fictrac on arch-based linux distros. It is fundamentally identical to the ubuntu one, just using pacman for the dependencies. It works on my machine, but has to be tested properly, I am sure i had already a lot of requirements preinstalled and experience may vary (especially given how varied arch distros are)